+++ app/Services/LinkContent/ProviderImage.php
@@ -1,18 +1,18 @@
 <?php
 
-namespace App\Services\ExtraContent;
+namespace App\Services\LinkContent;
 
 class ProviderImage extends BaseProvider
 {
     /** @var string $regex */
     public $regex = '/(?:https:\/\/)(?:.*)\.(?:png|jpg|jpeg|gif)/i';
 
-    public function check(): bool
+    public function canPreview(): bool
     {
         return preg_match($this->regex, $this->url) != false;
     }
 
-    public function get(): ?string
+    public function getPreview(): ?string
     {
         if (preg_match($this->regex, $this->url)) {
             return str_replace('{url}', $this->url, '<img src="{url}" class="img-fluid" />');
@@ -20,5 +20,4 @@ class ProviderImage extends BaseProvider
 
         return null;
     }
-
 }