+++ app/Services/LinkContent/ProviderSoundcloud.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace App\Services\ExtraContent;
+namespace App\Services\LinkContent;
 
 class ProviderSoundcloud extends BaseProvider
 {
-    public function check(): bool
+    public function canPreview(): bool
     {
         return preg_match('/https:\/\/(?:www\.)?soundcloud\.com\/(.*\/.*)/i', $this->url) != false;
     }
 
-    public function get(): ?string
+    public function getPreview(): ?string
     {
         $url = sprintf('http://soundcloud.com/oembed?format=json&url=%s&iframe=true', urlencode($this->url));