+++ app/Services/LinkContent/ProviderVideo.php
@@ -1,18 +1,18 @@
 <?php
 
-namespace App\Services\ExtraContent;
+namespace App\Services\LinkContent;
 
 class ProviderVideo extends BaseProvider
 {
     /** @var string $regex */
     public $regex = '/.*\.(mp4|mpg|mpeg)$/i';
 
-    public function check(): bool
+    public function canPreview(): bool
     {
         return preg_match($this->regex, $this->url) != false;
     }
 
-    public function get(): ?string
+    public function getPreview(): ?string
     {
         return str_replace('{url}', $this->url, '<video src="{url}" controls width="100%" height="auto"></video>');
     }