+++ database/migrations/2019_09_02_110000_add_archive_columns_to_links_table.php
@@ -1,0 +1,24 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddArchiveColumnsToLinksTable extends Migration
+{
+    public function up()
+    {
+        Schema::table('links', function (Blueprint $table) {
+            $table->string('archive')->nullable()->after('extra');
+            $table->renameColumn('extra', 'preview');
+        });
+    }
+
+    public function down()
+    {
+        Schema::table('links', function (Blueprint $table) {
+            $table->renameColumn('preview', 'extra');
+            $table->dropColumn('archive');
+        });
+    }
+}