+++ changelog.md
@@ -1,6 +1,15 @@
 # Unreleased
 
-## Fixes
+## Added
+
+- Link can now be archived using Puppeteer or Youtube-dl
+- Queue configuration
+
+## Changed
+
+- Link extra is now link preview
+
+## Fixed
 
 - Default settings now use the correct locale